Separate Editor (the view) and CodeMirror
I rename Editor to EditorAndPreview. Current Editor file contains
Editor header, body and foot components. This makes the project
structure easier to explore in Editor.js
